PR Feedback, save a seed for network keys in a file to allow easy changing.
