+++ crates/config/src/node.rs
@@ -12,6 +12,10 @@ use tracing::info;
 
 /// The filename to use when reading/writing the validator's BlsKey.
 pub const BLS_KEYFILE: &str = "bls.key";
+/// The filename to use when reading/writing the primary's network keys seed.
+pub const PRIMARY_NETWORK_SEED_FILE: &str = "primary.seed";
+/// The filename to use when reading/writing the network key seed used by all workers.
+pub const WORKER_NETWORK_SEED_FILE: &str = "worker.seed";
 
 /// Configuration for the Telcoin Network node.
 #[derive(Debug, Clone, Serialize, Deserialize)]
